test: Refactor test suite to run e2e test
Change-type: minor
