+++ index.js
@@ -1,59 +1,26 @@
 #!/usr/bin/env node
 
+const _ = require('lodash')
+const capitano = require('capitano')
+
 const {
   parse
 } = require('./lib/parser')
-const _ = require('lodash')
-const availableRules = require('./rules')
-const capitano = require('capitano')
-const fs = require('fs')
-const path = require('path')
-const storedErrors = []
-
-const parseRule = (command) => {
-  return _.map(command.split('-'), (word, index) => {
-    if (index !== 0) return word.charAt(0).toUpperCase().concat(word.slice(1))
-    return word
-  }).join('')
-}
-
-const runRules = (commit, rulesConfig, done) => {
-  _.forEach(rulesConfig, (configuration, rule) => {
-    // Return early if rule is not active
-    if (!configuration) return
-
-    const parsedRule = parseRule(rule)
-    if (_.isFunction(availableRules[parsedRule])) {
-      try {
-        availableRules[parsedRule](commit, configuration)
-      } catch (err) {
-        storedErrors.push(err)
-      }
-    } else {
-      throw new Error(`No available rule matching ${rule}`)
-    }
-  })
-  if (_.isEmpty(storedErrors)) {
-    done()
-  } else {
-    done(storedErrors)
-  }
-}
-
-const readConfig = (configPath) => {
-  const configFile = fs.readFileSync(configPath, 'utf8')
-  return JSON.parse(configFile)
-}
+const {
+  runRules,
+  readConfig,
+  getDefaultConfigPath
+} = require('./lib/runner')
 
 const parseAndRun = (params, options, done) => {
   let commit = {}
   try {
     commit = parse(params['commit-message'])
   } catch (err) {
-    return generateErrorReport([ err ])
+    return done([ err ])
   }
 
-  const configPath = path.join(__dirname, 'config.json')
+  const configPath = getDefaultConfigPath()
   const rulesConfig = readConfig(options.config || configPath)
   return runRules(commit, rulesConfig, done)
 }
