+++ test/e2e/commits/empty-prefix.yml
@@ -1,0 +1,1 @@
+throws: 'Expected prefix or subject but ":" found.'
