+++ test/e2e/commits/missing-footer.yml
@@ -1,0 +1,1 @@
+throws: "Could not parse commit: footers are mandatory and should be separated from the body with a newline"
