+++ test/e2e/commits/no-prefix.yml
@@ -1,0 +1,14 @@
+prefix: ''
+subject: 'simple title'
+fullTitle: 'simple title'
+body: body
+footers:
+- 'Footer: foo'
+message: |-
+  body
+
+  Footer: foo
+
+errors:
+- 'Signed-off-by must be the last tag appearing in the footers'
+- 'Each commit must contain the following footer: Signed-off-by: Full Name <email>'
