+++ test/e2e/commits/only-title.yml
@@ -1,0 +1,1 @@
+throws: "Expected newline between title and body but end of input found."
