+++ test/e2e/commits/title-with-newline.yml
@@ -1,0 +1,1 @@
+throws: "Expected body and footers but end of input found."
