+++ test/e2e/e2e.spec.js
@@ -1,15 +1,25 @@
 const ava = require('ava')
-const {
-  parse
-} = require('../../lib/parser')
 const fs = require('fs')
 const path = require('path')
 const yaml = require('js-yaml')
 const _ = require('lodash')
 
+const {
+  parse
+} = require('../../lib/parser')
+const {
+  runRules,
+  readConfig,
+  getDefaultConfigPath
+} = require('../../lib/runner')
+
 const tests = _.filter(fs.readdirSync(`${__dirname}/commits`), (filename) => {
   return path.extname(filename) !== '.yml'
 })
+
+const stripErrors = (parseResult) => {
+  return _.omit(parseResult, 'errors')
+}
 
 _.each(tests, (testName) => {
   const testCase = fs.readFileSync(`${__dirname}/commits/${testName}`, 'utf8')
@@ -18,14 +28,24 @@ _.each(tests, (testName) => {
   const expected = yaml.safeLoad(expectedFile)
 
   ava.test(testName, (test) => {
-    if (expected.error) {
+    if (expected.throws) {
       const error = test.throws(() => {
         parse(testCase)
       })
-      test.is(error.message, expected.error)
+      test.is(error.message, expected.throws)
     } else {
       const parseResult = parse(testCase)
-      test.deepEqual(parseResult, expected)
+      test.deepEqual(parseResult, stripErrors(expected))
+
+      const defaultConfig = readConfig(getDefaultConfigPath())
+      runRules(parseResult, defaultConfig, (errs) => {
+        const errorMessages = _.map(errs, (err) => {
+          return err.message
+        })
+        if (expected.errors) {
+          test.deepEqual(errorMessages, expected.errors)
+        }
+      })
     }
   })
 })
