--- test/parser/commits/empty-prefix.yml
@@ -1,1 +1,0 @@
-error: 'Expected prefix or subject but ":" found.'
