--- test/parser/commits/missing-footer.yml
@@ -1,1 +1,0 @@
-error: "Could not parse commit: footers are mandatory and should be separated from the body with a newline"
