--- test/parser/commits/no-prefix-no-body.yml
@@ -1,9 +1,0 @@
-prefix: ''
-subject: 'simple title'
-fullTitle: 'simple title'
-body: ''
-footers:
-- 'Footer: foo'
-message: |-
-
-  Footer: foo
