--- test/parser/commits/no-prefix.yml
@@ -1,10 +1,0 @@
-prefix: ''
-subject: 'simple title'
-fullTitle: 'simple title'
-body: body
-footers:
-- 'Footer: foo'
-message: |-
-  body
-
-  Footer: foo
