--- test/parser/commits/only-title.yml
@@ -1,1 +1,0 @@
-error: "Expected newline between title and body but end of input found."
