--- test/parser/commits/space-after-column.yml
@@ -1,11 +1,0 @@
-prefix: prefix
-subject: ' subject'
-fullTitle: 'prefix:  subject'
-body: body
-message: |-
-  body
-
-  Footer: foo
-footers:
-- 'Footer: foo'
-
