--- test/parser/commits/space-after-prefix.yml
@@ -1,10 +1,0 @@
-prefix: 'prefix '
-subject: subject
-fullTitle: 'prefix : subject'
-body: body
-footers:
-- 'Footer: foo'
-message: |-
-  body
-
-  Footer: foo
