--- test/parser/commits/title-with-newline.yml
@@ -1,1 +1,0 @@
-error: "Expected body and footers but end of input found."
